Modified bug fix in AzurePowerShell.ps1
Modified per advice from reviewer niadak. I opted to declare $__vstsAzPSInlineScriptPath = $null in line 28. That avoids the “cannot be retrieved” error occurring in the finally block if the called .ps1 script has executed "Set-PsDebug -Strict". (The variable $scriptType cannot be tested in the finally block, as it is removed earlier.)